backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
alembic==1.13.1
aiosqlite==0.19.0
apscheduler==3.10.4
python-multipart==0.0.6

